requirements.txt
requests==2.24.0
webscreenshot==2.9
python_telegram_bot==12.5.1
